Add VaultUnlocked support and update build/config
Register EconomyPlus as a VaultUnlocked-compatible economy provider and wire up build/config changes. Added VaultUnlocked provider classes (VaultHook, VaultUnlockedEconomy), added compileOnly dependency and CodeMC repo in build.gradle, and updated README with integration instructions. Bumped run/config.json version and enabled default mods, added/update runtime artifacts (logo, manifest, chunk/resource files, player/balance data and server logs). Miscellaneous code updates across API, commands and services to support the integration and related fixes.

+++ src/main/java/com/lucastudios/EconomyPlus/commands/BalCommand.java
@@ -17,6 +17,7 @@ import com.lucastudios.EconomyPlus.service.Utility;
 
 import javax.annotation.Nonnull;
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.text.NumberFormat;
 import java.util.HashMap;
 import java.util.Locale;
@@ -77,7 +78,7 @@ public final class BalCommand extends AbstractPlayerCommand {
             return;
         }
 
-        long balance = wallet.getBalance(currencyId);
+        BigDecimal balance = wallet.getBalance(currencyId);
         String formatted = formatBalance(currency, balance);
 
         Map<String, String> placeholders = new HashMap<>();
@@ -93,7 +94,7 @@ public final class BalCommand extends AbstractPlayerCommand {
         ctx.sendMessage(Message.raw(plugin.messages().format(messageKey, placeholders)));
     }
 
-    private String formatBalance(Currency currency, long balance) {
+    private String formatBalance(Currency currency, BigDecimal balance) {
         if (currency.decimals() == 0) {
             NumberFormat nf = NumberFormat.getIntegerInstance(Locale.ENGLISH);
             return nf.format(balance);
